Rework VectorUI class task 2 level d

+++ HomeworkStage14/src/by/itstep/oris/workspace/leveld/task02/view/VectorUI.java
@@ -3,9 +3,9 @@ package by.itstep.oris.workspace.leveld.task02.view;
 import java.util.Arrays;
 
 public class VectorUI {
-	public static void getResult(double sum, double... vector) {
+	public static void getResult(int sum, int... vector) {
 		System.out.println(String.format
-				("Sum of elements between first and last zero's in %s is: %.2f",
+				("Sum of elements between first and last zero's in %s is: %d",
 						 Arrays.toString(vector), sum));
 	}
 }
